feat: remove migration
