+++ parachain/runtime/litentry/src/lib.rs
@@ -99,7 +99,6 @@ pub mod asset_config;
 pub mod constants;
 pub mod precompiles;
 
-pub mod migration;
 pub mod weights;
 pub mod xcm_config;
 
@@ -157,7 +156,6 @@ pub type Executive = frame_executive::Executive<
 	// It was reverse order before.
 	// See the comment before collation related pallets too.
 	AllPalletsWithSystem,
-	migration::Migrations<Runtime>,
 >;
 
 impl fp_self_contained::SelfContainedCall for RuntimeCall {
